[confluence] Add In Progress submenu entry to TVSHOWS
This is a improvement in navigation and availability of
often used options.

It makes sense to me, to have this node available as submenu entry for
easy access from main menu and not have to go through non library
entries VIDEOS > In progress TV shows or Videos > Library > In Progress
TV shows to get to here.

Saves 10 seconds of your life.

There is probably a proper more elegant way to do this and if there is,
that would be preferable.